+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2020-04-05
 ### Added
 - Configuration file
 - Logging to file
@@ -24,4 +26,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial implementation of ssh transport
 - Basic listener
 
-[Unreleased]: https://gitlab.com/personal-assistant-bot/infrastructure/stomp-over-ssh/compare/v0.1.0...master
+[Unreleased]: https://gitlab.com/personal-assistant-bot/infrastructure/stomp-over-ssh/compare/v0.2.0...master
+[0.2.0]: https://gitlab.com/personal-assistant-bot/infrastructure/stomp-over-ssh/compare/v0.1.0...v0.2.0